[MajorVersion] File organization and module split (#9)
* [MajorVersion] Total file restructure and split files into modules
